src/ml-jimfleming--numerai/requirements.txt
Cython==0.27.3
matplotlib==2.1.1
numpy==1.14.0rc1
pandas==0.21.1
scikit-learn==0.19.1
scipy==1.0.0
tensorflow==1.4.1
tqdm==4.19.5
deap==1.2.2
TPOT==0.9.1
seaborn==0.8.1
jupyter==1.0.0
numexpr==2.6.4 # Required by SOMPY
xgboost==0.6a2
MulticoreTSNE==0.0.1.1
Pillow==5.0.0
